fix serverless-workflow-editor
