+++ packages/vscode-extension-serverless-workflow-editor/src/extension/serviceCatalog/serviceRegistry/index.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-export * from "./types";
 export * from "./ServiceRegistryStore";
 
 export function getServiceFileNameFromSwfServiceCatalogServiceId(
