--- packages/vscode-extension-serverless-workflow-editor/src/extension/serviceCatalog/serviceRegistry/types.ts
@@ -1,33 +1,0 @@
-/*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export enum AuthProviderType {
-  NONE = "none",
-  RH_ACCOUNT = "red-hat-account",
-  OIDC = "oidc",
-}
-
-export interface ServiceRegistrySettings {
-  name: string;
-  url: string;
-  authProvider: AuthProviderType;
-  authUrl?: string;
-  clientId?: string;
-}
-
-export interface ServiceRegistries {
-  registries: ServiceRegistrySettings[];
-}
